+++ test/linear_model/stepwise/both_selection_test.py
@@ -16,17 +16,20 @@ def data():
 def test_both_selection_ols(data):
     formula = "x1 ~ x2 + x3 + x4 + y"
     model = sm.OLS
-    final_formula = both_selection(formula=formula, data=data, model=model)
+    with pytest.deprecated_call():
+        final_formula = both_selection(formula=formula, data=data, model=model)
     assert final_formula == "x1 ~ x2 + x3 + x4", "Formula doesn't match expected"
 
 def test_both_selection_logit(data):
     formula = "y ~ x1 + x2 + x3 + x4"
     model = sm.Logit
-    final_formula = both_selection(formula=formula, data=data, model=model)
+    with pytest.deprecated_call():
+        final_formula = both_selection(formula=formula, data=data, model=model)
     assert final_formula == "y ~ x2 + x3 + x4", "Formula doesn't match expected"
     
 def test_both_selection_logisticregression(data):
     formula = "y ~ x1 + x2 + x3 + x4"
     model = LogisticRegression
-    final_formula = both_selection(formula=formula, data=data, model=model)
+    with pytest.deprecated_call():
+        final_formula = both_selection(formula=formula, data=data, model=model)
     assert final_formula == "y ~ x1 + x2 + x3 + x4", "Formula doesn't match expected"